+++ env-localdev
@@ -27,3 +27,7 @@ LDAP_FILTER_TMPL=(login=%%(login)s)
 DEFAULT_MYMAPS_ROLE=645
 SHORTENER_BASE_URL=http://localhost:8080/s/coco/../
 SHORTENER_ALLOWED_HOST=localhost
+# Guillaume's routing keys
+# Use them only for local dev
+ROUTING_GRAPHHOPPER_API_KEY=239fa42d-e0ef-482a-b1ca-68e84df1660b
+ROUTING_MAPQUEST_API_KEY=74Thu9D5G9eT1v8qEa9WBnx5zPcBejje
